\name{growthObjDeclineVar-class}
\Rdversion{1.1}
\docType{class}
\alias{growthObjDeclineVar-class}
%\alias{growth,numeric,numeric,numeric,growthObjDeclineVar-method}

\title{Class "growthObjDeclineVar"}
\description{
A class object description
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("growthObjDeclineVar", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{fit}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Methods}{
  \describe{
    \item{growth}{\code{signature(size = "numeric", sizeNext = "numeric", cov = "data.frame", growthObj = "growthObjDeclineVar")}: ... }
   \item{growthCum}{\code{signature(size = "numeric", sizeNext = "numeric", cov = "data.frame", growthObj = "growthObjDeclineVar")}: ... }
	 }
}
\author{
C. Jessica E. Metcalf, Sean M. McMahon, Roberto Salguero-Gomez, Eelke Jongejans & Cory Merow.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("growthObjDeclineVar")
}
\keyword{classes}
