\name{IPMpack-package}
\alias{IPMpack-package}
\alias{IPMpack}
\docType{package}
\title{
Construction and analysis of integral projection models and associated measures of population growth, structure, perturbations (sensitivities and elasticities), overall population dynamics, age-specific metrics, etc. 
}
\description{
IPM package, a series of analytical tools for building, diagnosing, and projecting populations models based on continuous and descrete vital rates.}
\details{
\tabular{ll}{
Package: \tab IPMpack\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2012-June-12\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Depends: \tab MCMCglmm, Matrix, MASS, nlme\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
The IPMpack team: C. Jessica E. Metcalf, Sean M. McMahon, Roberto Salguero-Gomez, Eelke Jongejans & Cory Merow.
Maintainer: ipmpack-users@lists.r-forge.r-project.org; ipmpack@gmail.com
User mailing list: https://lists.r-forge.r-project.org/mailman/listinfo/ipmpack-users 
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
%~~ in the R documentation directory ~~
\keyword{ package }
