\name{IPMpackNews}
\alias{IPMpackNews}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Show a text file containing package news and updates.
}
\description{
Learn about recent changes to IPMpack.
}
\usage{
IPMpackNews()
}

\author{
C. Jessica E. Metcalf, Sean M. McMahon, Roberto Salguero-Gomez, Eelke Jongejans & Cory Merow.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
IPMpackNews()
}



